Refactor push `views.py` into more appropriate modules
NHUB-548
